Add files via upload
This script is for large multi-FASTA file datasets. Define a character limit and it will delete all files that contain less than your character limit. This is script is useful if you download hundreds of FASTA files and only want to keep those with full plasmid or full genomes and need to weed out PCR amplicon sequences.